python_start_server/requirements.txt
redis
grpcio
protobuf
google